+++ app/src/main/java/com/makentoshe/habrachan/di/article/WebArticleFragmentModule.kt
@@ -1,7 +1,7 @@
 package com.makentoshe.habrachan.di.article
 
-import com.makentoshe.habrachan.model.article.JavaScriptInterface
-import com.makentoshe.habrachan.model.article.WebViewController
+import com.makentoshe.habrachan.model.article.web.JavaScriptInterface
+import com.makentoshe.habrachan.model.article.web.WebViewController
 import com.makentoshe.habrachan.view.article.WebArticleFragment
 import toothpick.config.Module
 import toothpick.ktp.binding.bind
@@ -11,6 +11,11 @@ class WebArticleFragmentModule(fragment: WebArticleFragment): Module() {
     init {
         val javascriptInterface = JavaScriptInterface()
         bind<JavaScriptInterface>().toInstance(javascriptInterface)
-        bind<WebViewController>().toInstance(WebViewController(fragment, javascriptInterface))
+        bind<WebViewController>().toInstance(
+            WebViewController(
+                fragment,
+                javascriptInterface
+            )
+        )
     }
 }
