+++ app/src/main/java/com/makentoshe/habrachan/model/article/web/html/ImageAddon.kt
@@ -1,24 +1,19 @@
-package com.makentoshe.habrachan.model.article.html
+package com.makentoshe.habrachan.model.article.web.html
 
 import org.jsoup.nodes.Element
 
 class ImageAddon : HtmlBuilder.Addon {
 
     override fun accept(body: Element) {
-        val sources = extractSources(body)
         val images = body.select("img")
         images.forEach { image ->
             fixUnclosedImgTag(image)
-            addClickListener(image, sources)
+            addClickListener(image)
         }
     }
 
-    private fun addClickListener(image: Element, sources: String) {
-        image.attr("onclick", "onImageClickedListener(this, \"$sources\")")
-    }
-
-    private fun extractSources(body: Element): String {
-        return body.select("img").joinToString(" ") { it.attr("src") }
+    private fun addClickListener(image: Element) {
+        image.attr("onclick", "onImageClickedListener(this)")
     }
 
     private fun fixUnclosedImgTag(image: Element) {
