+++ app/src/main/java/com/makentoshe/habrachan/model/article/web/html/TableAddon.kt
@@ -1,4 +1,4 @@
-package com.makentoshe.habrachan.model.article.html
+package com.makentoshe.habrachan.model.article.web.html
 
 import org.jsoup.nodes.Element
