+++ app/src/test/java/com/makentoshe/habrachan/view/acticle/WebArticleFragmentTest.kt
@@ -5,7 +5,7 @@ import com.makentoshe.habrachan.AppActivity
 import com.makentoshe.habrachan.di.article.ArticleFragmentScope
 import com.makentoshe.habrachan.di.article.WebArticleFragmentScope
 import com.makentoshe.habrachan.di.common.ApplicationScope
-import com.makentoshe.habrachan.model.article.JavaScriptInterface
+import com.makentoshe.habrachan.model.article.web.JavaScriptInterface
 import com.makentoshe.habrachan.model.article.WebArticleScreen
 import com.makentoshe.habrachan.view.article.ArticleFragment
 import com.makentoshe.habrachan.viewmodel.article.ArticleFragmentViewModel
